root conftest
# Rust tests and lints will be skipped. Delete if not intended.
[ci skip-rust]
