+++ BUILD
@@ -6,3 +6,5 @@ shell_sources(name="scripts", sources=["cargo", "pants"])
 # We use `BUILD_ROOT` to establish the build root, rather than `./pants`, per
 # https://github.com/pantsbuild/pants/pull/8105.
 files(name="files", sources=["BUILD_ROOT", ".gitignore", "pants.toml"])
+
+python_test_utils(name="test_utils")
